# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TODO Backend API",
        "description": "API para gerenciar tarefas (TODOs)"
    },
    "paths": {
        "/todos": {
            "get": {
                "summary": "Lista todas as tarefas",
                "tags": ["Tarefas"],
                "responses": {
                    "200": {
                        "description": "Lista de tarefas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Todo"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "summary": "Cria uma nova tarefa",
                "tags": ["Tarefas"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TodoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarefa criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Todo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/todos/{id}": {
            "put": {
                "summary": "Atualiza os status (done) de uma tarefa",
                "tags": ["Tarefas"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["done"],
                                "properties": {
                                    "done": {
                                        "type": "boolean",
                                        "default": false
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarefa atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Todo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                }
            },
            "delete": {
                "summary": "Deleta uma tarefa",
                "tags": ["Tarefas"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tarefa deletada"
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Todo": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "ID único da tarefa"
                    },
                    "title": {
                        "type": "string",
                        "description": "Título da tarefa"
                    },
                    "done": {
                        "type": "boolean",
                        "description": "Status de conclusão"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Data de criação"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Data de atualização"
                    }
                }
            },
            "TodoInput": {
                "type": "object",
                "required": ["title"],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Título da tarefa"
                    },
                    "done": {
                        "type": "boolean",
                        "default": false
                    }
                }
            }
        }
    }
}